% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algoleverage.R
\name{algoleverage}
\alias{algoleverage}
\title{Subsampling regression methods}
\usage{
algoleverage(X, y, samplesize, Type = "Uniform", seed)
}
\arguments{
\item{X}{The covariates matrix of size \code{n} by \code{p}.}

\item{y}{The response vector with length matching to the row number of \code{X}.}

\item{samplesize}{Number of subsamples to choose for regression.}

\item{Type}{The type of subsampling rows. "Lev" & "Uniform", the default is "Uniform"}

\item{seed}{set the seed}
}
\value{
The coefficient vector \code{beta}.
}
\description{
Implements algorithmic leveraging for linear regression using uniform or
     leverage score based subsampling of rows, i.e, fit \code{y = X beta + E}.
}
